requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
idna==2.10
gunicorn==19.9.0
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.5
python3-openid==3.2.0
python-decouple==3.3
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
